ssp: add APIs for Stack Smashing Protection
Compiling with any of the -fstack-protector* flags requires the
__stack_chk_guard data import (which needs to be initialized) and the
__stack_chk_fail{,_local} functions.  While GCC's own libssp can provide
these, it is better that we provide these ourselves.  The header is from
NetBSD but the implementation is custom due to NetBSD's implementation
being OS-specific.

Note that this does require building gcc with --disable-libssp (as with
other targets which provide their own SSP implementation) and
gcc_cv_libc_provides_ssp=yes (which should be at least conditionally fixed
in gcc) to avoid gcc trying to link with -lssp.

Signed-off-by: Yaakov Selkowitz <yselkowi@redhat.com>