ssp: add Object Size Checking for string.h functions
The Object Size Checking (-D_FORTIFY_SOURCE=*) functionality provides
wrappers around functions suspectible to buffer overflows.  While
independent from Stack Smashing Protection (-fstack-protector*), they
are often used and implemented together.

While GCC also provides an implementation in libssp, it is completely
broken (CVE-2016-4973, RHBZ#1324759) and seemingly unfixable, as there
is no reliable way for a preprocessor macro to trigger a link flag.
Therefore, adding this here is necessary to make it work.

Note that this does require building gcc with --disable-libssp, as with
other targets which provide their own implementation, as the headers from
its libssp would clash with this one.

The implementation is from NetBSD, with the addition of mempcpy (a GNU
extension) for parity with libssp and glibc.

Signed-off-by: Yaakov Selkowitz <yselkowi@redhat.com>